portalicious: implement "auth strategies"
