+++ interfaces/Portalicious/src/app/components/page-layout/components/header/header.component.html
@@ -21,7 +21,7 @@
         label="Account"
         i18n-label="Top-right user-menu@@header-user-menu"
         [icon]="'pi pi-user'"
-        [menuItems]="userMenuOptions"
+        [menuItems]="userMenuOptions()"
         class="[&_.p-button-label]:duration-0 [&_button:not(:hover)]:text-white"
         [plain]="true"
         [text]="true"
