+++ interfaces/Portalicious/src/app/guards/auth-capabilities.guard.ts
@@ -1,0 +1,17 @@
+import { inject } from '@angular/core';
+import { CanActivateFn, Router } from '@angular/router';
+
+import { AuthService } from '~/services/auth.service';
+
+export const authCapabilitiesGuard =
+  (capabilityChecker: (authService: AuthService) => boolean): CanActivateFn =>
+  () => {
+    const authService = inject(AuthService);
+
+    if (capabilityChecker(authService)) {
+      return true;
+    }
+
+    const router = inject(Router);
+    return router.navigate(['/']);
+  };
