+++ interfaces/Portalicious/src/app/pages/login/login.page.ts
@@ -1,49 +1,23 @@
-import { NgOptimizedImage } from '@angular/common';
-import {
-  ChangeDetectionStrategy,
-  Component,
-  computed,
-  inject,
-} from '@angular/core';
-import {
-  FormControl,
-  FormGroup,
-  ReactiveFormsModule,
-  Validators,
-} from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
+import { NgComponentOutlet, NgOptimizedImage } from '@angular/common';
+import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 
-import { injectMutation } from '@tanstack/angular-query-experimental';
-import { AutoFocusModule } from 'primeng/autofocus';
-import { InputTextModule } from 'primeng/inputtext';
-import { PasswordModule } from 'primeng/password';
 import { ToolbarModule } from 'primeng/toolbar';
 
-import { FormDefaultComponent } from '~/components/form/form-default.component';
-import { FormFieldWrapperComponent } from '~/components/form-field-wrapper/form-field-wrapper.component';
 import { LanguageSwitcherComponent } from '~/components/language-switcher/language-switcher.component';
 import { LogoComponent } from '~/components/logo/logo.component';
 import { CookieBannerComponent } from '~/pages/login/components/cookie-banner/cookie-banner.component';
 import { AuthService } from '~/services/auth.service';
-import { generateFieldErrors } from '~/utils/form-validation';
-
-type LoginFormGroup = (typeof LoginPageComponent)['prototype']['formGroup'];
 
 @Component({
   selector: 'app-login',
   standalone: true,
   imports: [
-    InputTextModule,
-    PasswordModule,
-    AutoFocusModule,
     LogoComponent,
     ToolbarModule,
     NgOptimizedImage,
     CookieBannerComponent,
     LanguageSwitcherComponent,
-    FormDefaultComponent,
-    ReactiveFormsModule,
-    FormFieldWrapperComponent,
+    NgComponentOutlet,
   ],
   templateUrl: './login.page.html',
   styles: ``,
@@ -51,57 +25,6 @@ type LoginFormGroup = (typeof LoginPageComponent)['prototype']['formGroup'];
 })
 export class LoginPageComponent {
   private authService = inject(AuthService);
-  private route = inject(ActivatedRoute);
-  private router = inject(Router);
 
-  private returnUrl = computed(() => {
-    const returnUrl: unknown = this.route.snapshot.queryParams.returnUrl;
-    if (typeof returnUrl !== 'string') {
-      return undefined;
-    }
-    return returnUrl;
-  });
-
-  formGroup = new FormGroup({
-    email: new FormControl('', {
-      nonNullable: true,
-      // eslint-disable-next-line @typescript-eslint/unbound-method
-      validators: [Validators.required, Validators.email],
-    }),
-    password: new FormControl('', {
-      nonNullable: true,
-      // eslint-disable-next-line @typescript-eslint/unbound-method
-      validators: [Validators.required],
-    }),
-  });
-
-  formFieldErrors = generateFieldErrors<LoginFormGroup>(this.formGroup, {
-    email: (control) => {
-      if (!control.invalid) {
-        return;
-      }
-      return $localize`Enter a valid email address`;
-    },
-    password: (control) => {
-      if (!control.invalid) {
-        return;
-      }
-      return $localize`Enter your password`;
-    },
-  });
-
-  loginMutation = injectMutation(() => ({
-    mutationFn: ({
-      email,
-      password,
-    }: ReturnType<LoginFormGroup['getRawValue']>) =>
-      this.authService.login({ username: email, password }),
-    onSuccess: () => {
-      const returnUrl = this.returnUrl();
-      if (returnUrl) {
-        return this.router.navigate([returnUrl]);
-      }
-      return this.router.navigate(['/']);
-    },
-  }));
+  LoginComponent = this.authService.LoginComponent;
 }
