+++ interfaces/Portalicious/src/app/services/auth/strategies/basic-auth/basic-auth.change-password.component.html
@@ -1,0 +1,44 @@
+<app-form-default
+  [formGroup]="formGroup"
+  submitButtonText="Change password"
+  i18n-submitButtonText
+  [mutation]="changePasswordMutation"
+>
+  <app-form-field-wrapper
+    label="Current Password"
+    i18n-label
+    [errorMessage]="formFieldErrors()('currentPassword')"
+  >
+    <p-password
+      formControlName="currentPassword"
+      autocomplete="current-password"
+      [toggleMask]="true"
+      [feedback]="false"
+      pAutoFocus
+      autofocus
+    />
+  </app-form-field-wrapper>
+  <app-form-field-wrapper
+    label="New Password"
+    i18n-label
+    [errorMessage]="formFieldErrors()('newPassword')"
+  >
+    <p-password
+      formControlName="newPassword"
+      autocomplete="new-password"
+      [toggleMask]="true"
+      [feedback]="false"
+    />
+  </app-form-field-wrapper>
+  <app-form-field-wrapper
+    label="Confirm Password"
+    i18n-label
+    [errorMessage]="formFieldErrors()('confirmPassword')"
+  >
+    <p-password
+      formControlName="confirmPassword"
+      [toggleMask]="true"
+      [feedback]="false"
+    />
+  </app-form-field-wrapper>
+</app-form-default>
