+++ interfaces/Portalicious/src/app/services/auth/strategies/msal-auth/msal-auth.strategy.ts
@@ -1,0 +1,10 @@
+import { Injectable } from '@angular/core';
+
+import { BasicAuthStrategy } from '~/services/auth/strategies/basic-auth/basic-auth.strategy';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class MsalAuthStrategy extends BasicAuthStrategy {
+  // TODO: Remove "extends BasicAuthStrategy" and implement MSAL authentication
+}
